Peak frequency code added
Calculated the peak frequency not sure whether it's working fully,
requires further testing and an implementation with LEDs. The code for
calculating the FFT that was present in the AudioAdjuster example code
was used to calculate the FFT and now using a snippet of code (that
allows the calculation of the peak) from an example provided by the
DSPlib (CE018_FFT_DSPlib).