Assignment2_Version2
Added a sort of recording function, switch 1 is pressed and recording
starts and the frctAudioOut array was enlarged to x5 the FRAME_SIZE as
to allow for more storage, also the memory attribute was changed from
"xmemory" to "ymemory" to accommodate for the additional memory. Switch
1 can be pressed again to reset the frctAudioOut array.

Might need to consider adding a delay of some sort, the sound that is
recorded is very short.